update util, add some more test datasets

+++ src/spatialjoin/Sweeper.cpp
@@ -3820,10 +3820,10 @@ double Sweeper::getMaxScaleFactor(const I32Point& p) const {
 // _____________________________________________________________________________
 double Sweeper::meterDist(const I32Point& p1, const I32Point& p2,
                           double maxDist) {
-  auto fp1 = FPoint{(p1.getX() * 1.0) / (PREC * 1.0),
-                    (p1.getY() * 1.0) / (PREC * 1.0)};
-  auto fp2 = FPoint{(p2.getX() * 1.0) / (PREC * 1.0),
-                    (p2.getY() * 1.0) / (PREC * 1.0)};
+  auto fp1 = FPoint{static_cast<float>((p1.getX() * 1.0) / (PREC * 1.0)),
+                    static_cast<float>((p1.getY() * 1.0) / (PREC * 1.0))};
+  auto fp2 = FPoint{static_cast<float>((p2.getX() * 1.0) / (PREC * 1.0)),
+                    static_cast<float>((p2.getY() * 1.0) / (PREC * 1.0))};
 
   double dX = fp2.getX() - fp1.getX();
   double dY = fp2.getY() - fp1.getY();
